net/ice: fix memory leaks in error handlings
The function release 'pkt_buf' and 'msk_buf' when successful. In some
error handlings, 'pkt_buf' and 'msk_buf' were not freed leads to memory
leaks.
To fix it without introducing code redundancy, add and make use of an
'error' lable to do the clean ups for 'pkt_buf' and 'msk_buf'.

Fixes: 1b9c68120a1c50 ("net/ice: enable protocol agnostic flow offloading in RSS")

Signed-off-by: Weiguo Li <liwg06@foxmail.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>