raw/ifpga/base: fix memory leaks in error handlings
1) in ifpga_enumerate.c:
The memory 'feature' is stored by TAILQ_INSERT_TAIL() at the end of the
function. When function returned early in error handling, 'feature' is
not released and leads to a memory leak.

2) in opae_eth_group.c and opae_i2c.c
These function return 'dev' when success and return NULL when validation
failed or some error occur. In the latter case 'dev' is not released and
leads to a memory leak.

Fixes: 56bb54ea1bdf ("raw/ifpga/base: add Intel FPGA OPAE share code")
Fixes: 12f92a513a13 ("raw/ifpga/base: fix retimer link status")
Fixes: 15d21c851028 ("raw/ifpga/base: add I2C and at24 EEPROM driver")

Signed-off-by: Weiguo Li <liwg06@foxmail.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>