common/cnxk: support for CPT second pass flow rules
Added support to create flow rules to match packets
from CPT(second pass packets). With this change, ingress
rules will be created with bits 10 and 11 of channel field
in the MCAM ignored by default. For rules specific to
second pass packets, the CPT channel bits will be set
in the MCAM.

Signed-off-by: Satheesh Paul <psatheesh@marvell.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>