allow using standard ar from the build ini files instead of 'ar' string
Hi,

I've attached a patch, from Gentoo, which uses the *ar* binary passed to
meson when available, instead of 'ar', which may not be available, for
instance when cross compiling, or having multiple gcc versions present,
like in Gentoo.

This should not have any regressions, as when the binary is not
available it uses the same logic as the original.

Aisha

Signed-off-by: 0-day Robot <robot@bytheb.org>