devtools: use libabigail rule for mlx glue drivers
Convert the existing exception in the ABI script into a libabigail
suppression rule.

Note: file_name_regexp could be used to achive the same with versions of
libabigail < 1.7 but soname_regexp has been preferred here since it is
already used with a recent change on common/mlx5.

While at it, fix indent from a recent change.

Signed-off-by: David Marchand <david.marchand@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>