app/testpmd: add help messages for multi-process
This patch adds help messages for multi-process.
--num-procs <N>, where N is the total number of symmetric_mp instances
that will be run side-by-side to perform packet processing. This parameter
is used to configure the appropriate number of receive queues on each
network port.

--proc-id <n>, where n is a numeric value in the range 0 <= n < N (number
of processes, specified above). This identifies which symmetric_mp instance
is being run, so that each process can read a unique receive queue on each
network port.

Fixes: a550baf24af9 ("app/testpmd: support multi-process")
Cc: stable@dpdk.org

Signed-off-by: Huisong Li <lihuisong@huawei.com>
Signed-off-by: Min Hu (Connor) <humin29@huawei.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>