devtools/cocci: change boolean negation to bitwise negation
Combining boolean operator with bitwise operator is suspicious. When
this happens, it has a chance that the bitwise negation operation is
mistakenly written as a boolean negation operation. This script is
used to find this kind of problems.

example:
	if (x & !BITS_MASK)
changed to:
	if (x & ~BITS_MASK)

The idea came from a demo script in coccinelle website:
https://coccinelle.gitlabpages.inria.fr/website/rules/notand.html

Signed-off-by: Weiguo Li <liwg06@foxmail.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>