net/iavf: replace SMP barrier with thread fence
Replace the SMP barrier with atomic thread fence for iavf hw ring scan
in the bulk Rx path.

This patch introduces a change to the iavf driver that was already added
to the i40e driver [1] as part of the adoption of the use of compiler
atomics.

[1] https://patches.dpdk.org/project/dpdk/patch/20210706065404.25137-3-
joyce.kong@arm.com/

Signed-off-by: Kathleen Capella <kathleen.capella@arm.com>
Reviewed-by: Joyce Kong <joyce.kong@arm.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>