build: encapsulate grep for local_miss_maps in a function
check-symbols-maps.sh runs with the -e option to exit as soon as
there is an error. Currently the build is failing as the search of
grep -L is empty (which is the case for all version.map files as
they have the local symbol) which causes the script to return
with an error.
This patch encapsulates the grep call in a function, so that only the function
errors out and the build is successful.

Fixes: b403498e14 ("build: hide local symbols in shared libraries")
Cc: stable@dpdk.org
Signed-off-by: Usama Arif <usama.arif@bytedance.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>